+++ lib/Conversion/TcpToArith/TcpToArith.cpp
@@ -9,8 +9,8 @@
 
 #include "Conversion/TcpToArith/TcpToArith.h"
 
-#include "IR/TcpDialect.h"
-#include "IR/TcpOps.h"
+#include "Dialect/IR/TcpDialect.h"
+#include "Dialect/IR/TcpOps.h"
 
 #include "../PassDetail.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
