+++ lib/Conversion/TcpToLinalg/Misc.cpp
@@ -10,8 +10,8 @@
 #include "Conversion/TcpToLinalg/TcpToLinalg.h"
 
 #include "../PassDetail.h"
-#include "IR/TcpDialect.h"
-#include "IR/TcpOps.h"
+#include "Dialect/IR/TcpDialect.h"
+#include "Dialect/IR/TcpOps.h"
 #include "PopulatePatterns.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
