+++ lib/Conversion/TorchToTcp/TorchToTcp.cpp
@@ -10,8 +10,8 @@
 #include "Conversion/TorchToTcp/TorchToTcp.h"
 
 #include "../PassDetail.h"
-#include "IR/TcpDialect.h"
-#include "IR/TcpOps.h"
+#include "Dialect/IR/TcpDialect.h"
+#include "Dialect/IR/TcpOps.h"
 #include "PopulatePatterns.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/Tensor/IR/Tensor.h"
