+++ lib/Dialect/IR/TcpOps.cpp
@@ -7,7 +7,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "IR/TcpOps.h"
+#include "Dialect/IR/TcpOps.h"
 
 #include "mlir/IR/Builders.h"
 #include "mlir/IR/OpImplementation.h"
@@ -16,7 +16,7 @@
 #include "mlir/IR/Value.h"
 
 #define GET_OP_CLASSES
-#include "IR/TcpOps.cpp.inc"
+#include "Dialect/IR/TcpOps.cpp.inc"
 
 namespace mlir::tcp {
